FileParsers reorganization, tidy up comments and moved summon parser from summons.c to init-monster.c
