+++ src/mon-summon.h
@@ -47,6 +47,8 @@ int summon_name_to_idx(const char *name);
 int summon_fallback_type(int summon_type);
 int summon_message_type(int summon_type);
 const char *summon_desc(int type);
+void create_summons(struct summon *parsed_summons);
+void free_summons(void);
 int summon_specific(struct loc grid, int lev, int type, bool delay, bool call);
 struct monster_race *select_shape(struct monster *mon, int type);
